requirements/development.txt
# requirements/development.txt
-r common.txt
Werkzeug==0.9.6

django-extensions==1.3.10
django-rosetta==0.7.4
django-debug-toolbar==1.2.1

transifex-client==0.10
